Refactor Auth Controller
